API modifications

+++ EveHack/GarduinoV2/main/models.py
@@ -14,11 +14,11 @@ class PlantData(models.Model):
 
     # All required data from plant
     name = models.CharField(max_length=50, default="Not set")
-    temp = models.IntegerField(default=0)
-    humidity = models.IntegerField(default=0)
+    temp = models.FloatField(default=0.0)
+    humidity = models.FloatField(default=0.0)
     next_spray = models.TimeField()
-    fertilizer_level = models.IntegerField(default=0)
-    led_intensity = models.IntegerField(default=0)
+    fertilizer_level = models.FloatField(default=0.0)
+    led_intensity = models.FloatField(default=0.0)
     time_to_sundown = models.TimeField()
     status = models.CharField(max_length=1, choices=PLANT_STATUS, default="N")
